+++ src/selection/useSelection.ts
@@ -266,7 +266,10 @@ export const useSelection = ({
           [data.id],
           pathSelectionType
         );
-        ref.current?.centerGraph([data.id, ...adjacents]);
+
+        ref.current.fitNodesInView([data.id, ...adjacents], {
+          fitOnlyIfNodesNotInView: true
+        });
       }
     },
     [
@@ -349,7 +352,7 @@ export const useSelection = ({
             throw new Error('No ref found for the graph canvas.');
           }
 
-          ref.current?.centerGraph();
+          ref.current.fitNodesInView([], { fitOnlyIfNodesNotInView: true });
         }
       }
     },