+++ src/symbols/Node.tsx
@@ -11,7 +11,6 @@ import { animationConfig } from '../utils';
 import { useSpring, a } from '@react-spring/three';
 import { Sphere } from './nodes/Sphere';
 import { Label } from './Label';
-import { Ring } from './Ring';
 import {
   NodeContextMenuProps,
   ContextMenuEvent,
@@ -25,6 +24,7 @@ import { useStore } from '../store';
 import { useDrag } from '../utils/useDrag';
 import { Icon } from './nodes';
 import { useHoverIntent } from '../utils/useHoverIntent';
+import { ThreeEvent } from '@react-three/fiber';
 
 export interface NodeProps {
   /**
@@ -70,22 +70,35 @@ export interface NodeProps {
   /**
    * The function to call when the pointer is over the node.
    */
-  onPointerOver?: (node: InternalGraphNode) => void;
+  onPointerOver?: (
+    node: InternalGraphNode,
+    event: ThreeEvent<PointerEvent>
+  ) => void;
 
   /**
    * The function to call when the pointer is out of the node.
    */
-  onPointerOut?: (node: InternalGraphNode) => void;
+  onPointerOut?: (
+    node: InternalGraphNode,
+    event: ThreeEvent<PointerEvent>
+  ) => void;
 
   /**
    * The function to call when the node is clicked.
    */
-  onClick?: (node: InternalGraphNode, props?: CollapseProps) => void;
+  onClick?: (
+    node: InternalGraphNode,
+    props?: CollapseProps,
+    event?: ThreeEvent<MouseEvent>
+  ) => void;
 
   /**
    * The function to call when the node is double clicked.
    */
-  onDoubleClick?: (node: InternalGraphNode) => void;
+  onDoubleClick?: (
+    node: InternalGraphNode,
+    event: ThreeEvent<MouseEvent>
+  ) => void;
 
   /**
    * The function to call when the node is right clicked.
@@ -203,12 +216,10 @@ export const Node: FC<NodeProps> = ({
     onDragStart: () => {
       setDraggingId(id);
       setActive(true);
-      cameraControls.controls.enabled = false;
     },
     onDragEnd: () => {
       setDraggingId(null);
       setActive(false);
-      cameraControls.controls.enabled = true;
       onDragged?.(node);
     }
   });
@@ -227,13 +238,15 @@ export const Node: FC<NodeProps> = ({
 
   const { pointerOver, pointerOut } = useHoverIntent({
     disabled: disabled || isDragging,
-    onPointerOver: () => {
+    onPointerOver: (event: ThreeEvent<PointerEvent>) => {
+      cameraControls.controls.truckSpeed = 0;
       setActive(true);
-      onPointerOver?.(node);
+      onPointerOver?.(node, event);
     },
-    onPointerOut: () => {
+    onPointerOut: (event: ThreeEvent<PointerEvent>) => {
+      cameraControls.controls.truckSpeed = 2.0;
       setActive(false);
-      onPointerOut?.(node);
+      onPointerOut?.(node, event);
     }
   });
 
@@ -247,6 +260,7 @@ export const Node: FC<NodeProps> = ({
           active: combinedActiveState,
           opacity: selectionOpacity,
           animated,
+          selected: isSelected,
           node
         })
       ) : (
@@ -261,6 +275,7 @@ export const Node: FC<NodeProps> = ({
               color={color}
               node={node}
               active={combinedActiveState}
+              selected={isSelected}
             />
           ) : (
             <Sphere
@@ -271,6 +286,7 @@ export const Node: FC<NodeProps> = ({
               color={color}
               node={node}
               active={combinedActiveState}
+              selected={isSelected}
             />
           )}
         </>
@@ -283,6 +299,7 @@ export const Node: FC<NodeProps> = ({
       combinedActiveState,
       selectionOpacity,
       animated,
+      isSelected,
       node
     ]
   );
@@ -365,22 +382,27 @@ export const Node: FC<NodeProps> = ({
 
   return (
     <a.group
+      renderOrder={1}
       userData={{ id, type: 'node' }}
       ref={group}
       position={nodePosition as any}
       onPointerOver={pointerOver}
       onPointerOut={pointerOut}
-      onClick={() => {
+      onClick={(event: ThreeEvent<MouseEvent>) => {
         if (!disabled && !isDragging) {
-          onClick?.(node, {
-            canCollapse,
-            isCollapsed
-          });
+          onClick?.(
+            node,
+            {
+              canCollapse,
+              isCollapsed
+            },
+            event
+          );
         }
       }}
-      onDoubleClick={() => {
+      onDoubleClick={(event: ThreeEvent<MouseEvent>) => {
         if (!disabled && !isDragging) {
-          onDoubleClick?.(node);
+          onDoubleClick?.(node, event);
         }
       }}
       onContextMenu={() => {
@@ -396,12 +418,6 @@ export const Node: FC<NodeProps> = ({
       {...(bind() as any)}
     >
       {nodeComponent}
-      <Ring
-        opacity={isSelected ? 0.5 : 0}
-        size={nodeSize}
-        animated={animated}
-        color={isSelected || active ? theme.ring.activeFill : theme.ring.fill}
-      />
       {menuComponent}
       {labelComponent}
     </a.group>